+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     `ConnectionType` enum.
   - [Work in progress] Model transports:
     - `polling`: `PollingTransport`
+    - `websocket`: `WebSocketTransport`
 - [Work in progress] Define the `Server` class:
   - [Work in progress] Define the `ServerConfiguration` class to contain the
     available configuration options for the server.